requirements.txt
# fastapi==0.115.0
# uvicorn[standard]==0.32.0
# python-dotenv==1.0.1
# httpx==0.27.2
# pydantic==2.9.2
# google-generativeai==0.8.3


fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.1
google-generativeai==0.3.1
